refactor: update file import path for openFiles utility and create new file-picker module
